Empty Ruby read buffer on transport open
This empties the read buffer when the transport is opened in the framed
transport for Ruby.

The commit should fix that when an error occured (for instance when no fields
are set in a union) we got data in the buffer in the subsequent requests. This
ended up corrupting all requests after the errored one.